rename chat_to_link to get_chat_link
